feat: create unified GenericListPanel and fix TUI navigation
- Replace MainPanel and SecondaryPanel with unified GenericListPanel component
- Fix NavigationContext error by restructuring component hierarchy
- Fix flickering issue by optimizing navigation priority and update triggers
- Simplify FirstRunWizard to work without GenericListPanel (temporary)
- Remove excessive debug logging that was causing performance issues
- Ensure wizard completes successfully and saves configuration

The TUI now loads properly and navigation works without flickering.
Next step: Update wizard to use FilePickerListItem for folder selection.
